requirements.txt
# Core LLM frameworks
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Security tools integration
python-nmap>=0.7.0
searchsploit>=0.1.0
py-exploitdb>=1.0.0
nvdlib>=0.4.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.66.0

# For sandbox execution
docker>=6.1.0

# Optional: Local LLM support
ollama>=0.3.0
flask>=2.3.0
flask-cors>=4.0.0